+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Increased the engine version to 5.10.
 - String terminator is not required anymore in genedits context functions dealing with arrays of characters
 - Fixed an issue in GEN_MATCH genedits context function.
+- Fixed logic that combines translated error messages on RuleFailure when there are several messages.
 
 **Version 1.20**